+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ### Devop
 
+* Move tokens to state [#2898](https://github.com/MyEtherWallet/MyEtherWallet/pull/2898)
+* Refactor empty balance block, hide elemnents on network switches [#2890](https://github.com/MyEtherWallet/MyEtherWallet/pull/2890)
+* Mobile Navigation bar implementationn for wallet, repalce mdi-text icon to svg [#2895](https://github.com/MyEtherWallet/MyEtherWallet/pull/2895)
+* Refactor access wallet, update ui to make it consistent \[#2846] (https://github.com/MyEtherWallet/MyEtherWallet/pull/2846)
 * Fix langing page header [#2886](https://github.com/MyEtherWallet/MyEtherWallet/pull/2886)
 * Update swap provider list [#2871](https://github.com/MyEtherWallet/MyEtherWallet/pull/2871)
 * Settings [#2835h](https://github.com/MyEtherWallet/MyEtherWallet/pull/2835)